+++ src/backend/npm.rs
@@ -77,7 +77,7 @@ impl Backend for NPMBackend {
                 .arg(tv.install_path())
                 .arg("--global")
                 .arg("--trust")
-                .with_pr(ctx.pr.as_ref())
+                .with_pr(&ctx.pr)
                 .envs(ctx.ts.env_with_path(&config)?)
                 .env("BUN_INSTALL_GLOBAL_DIR", tv.install_path())
                 .env("BUN_INSTALL_BIN", tv.install_path().join("bin"))
@@ -91,7 +91,7 @@ impl Backend for NPMBackend {
                 .arg(format!("{}@{}", self.tool_name(), tv.version))
                 .arg("--prefix")
                 .arg(tv.install_path())
-                .with_pr(ctx.pr.as_ref())
+                .with_pr(&ctx.pr)
                 .envs(ctx.ts.env_with_path(&config)?)
                 .prepend_path(ctx.ts.list_paths())?
                 .prepend_path(self.dependency_toolset()?.list_paths())?